Separate all JSX elements into their own compoment
