+++ components/Block.tsx
@@ -1,0 +1,9 @@
+import styles from "../styles/MathProblem.module.scss";
+
+export const Block = ({ children }: { children: number }) => {
+	return (
+		<article className={styles.block}>
+			<p className={styles.number}>{children}</p>
+		</article>
+	);
+};
